requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python
Pillow
tensorflow
scikit-learn
flask
fastapi
gradio
streamlit
torch
torchvision